refactor(cli): reuse defra-p2p-adapter for P2P HTTP adapters
Collapse duplicated CLI libp2p/iroh adapters, version syncers, and
document pushers into thin re-exports of the shared crate. Port CLI
explicit-replay validation and capability-change replay selection into
defra-p2p-adapter so both CLI and embedded keep the stricter semantics.
