+++ crates/cli/src/lib.rs
@@ -27,15 +27,15 @@ pub(crate) mod nac_adapter;
 #[allow(dead_code)]
 pub(crate) mod p2p_adapter;
 #[allow(dead_code)]
-pub(crate) mod p2p_adapter_helpers;
-#[allow(dead_code)]
 pub(crate) mod p2p_collection_lookup;
 #[allow(dead_code)]
 pub(crate) mod p2p_doc_pusher;
 pub(crate) mod schema_adapter;
 pub(crate) mod sourcehub_acp_adapter;
+#[cfg(feature = "iroh")]
 #[allow(dead_code)]
 pub(crate) mod transport_doc_pusher;
+#[cfg(feature = "iroh")]
 #[allow(dead_code)]
 pub(crate) mod transport_version_syncer;
 pub(crate) mod txn_adapter;
