+++ crates/p2p-adapter/src/lib.rs
@@ -59,6 +59,28 @@ where
         .collect()
 }
 
+/// Collections that need an initial document replay when adding a replicator.
+///
+/// Replays collections that are newly authorized, plus existing collections whose
+/// explicit replay capability changed (so a previously invalid authorizer can be
+/// corrected without requiring a remove/add cycle).
+pub fn collections_requiring_replay(
+    effective_collections: &[String],
+    collection_cids: &[String],
+    existing_collection_ids: &std::collections::HashSet<String>,
+    collections_with_changed_capabilities: &std::collections::HashSet<String>,
+) -> Vec<String> {
+    effective_collections
+        .iter()
+        .zip(collection_cids.iter())
+        .filter(|(_, cid)| {
+            !existing_collection_ids.contains(*cid)
+                || collections_with_changed_capabilities.contains(*cid)
+        })
+        .map(|(name, _)| name.clone())
+        .collect()
+}
+
 /// Overlay persisted peer metadata onto live replicator authorization state.
 ///
 /// The live registry/transport is authoritative for which peers and collections
@@ -112,8 +134,11 @@ pub fn merge_live_replicators_with_persisted_metadata(
 
 #[cfg(test)]
 mod resolve_remove_collections_tests {
-    use super::{merge_live_replicators_with_persisted_metadata, resolve_remove_collections};
-    use std::collections::HashMap;
+    use super::{
+        collections_requiring_replay, merge_live_replicators_with_persisted_metadata,
+        resolve_remove_collections,
+    };
+    use std::collections::{HashMap, HashSet};
 
     fn resolver(map: HashMap<&'static str, &'static str>) -> impl Fn(&str) -> Option<String> {
         move |name| map.get(name).map(|cid| cid.to_string())
@@ -138,6 +163,40 @@ mod resolve_remove_collections_tests {
         let map = HashMap::from([("AgentDoc", "bafyCID")]);
         let out = resolve_remove_collections(Vec::new(), resolver(map));
         assert!(out.is_empty());
+    }
+
+    #[test]
+    fn collections_requiring_replay_replays_existing_collection_when_capability_changes() {
+        let effective_collections = vec!["User".to_string()];
+        let collection_cids = vec!["cid-user".to_string()];
+        let existing_collection_ids = HashSet::from(["cid-user".to_string()]);
+        let changed_capabilities = HashSet::from(["cid-user".to_string()]);
+
+        let replay_collections = collections_requiring_replay(
+            &effective_collections,
+            &collection_cids,
+            &existing_collection_ids,
+            &changed_capabilities,
+        );
+
+        assert_eq!(replay_collections, vec!["User".to_string()]);
+    }
+
+    #[test]
+    fn collections_requiring_replay_skips_existing_collection_when_capability_matches() {
+        let effective_collections = vec!["User".to_string()];
+        let collection_cids = vec!["cid-user".to_string()];
+        let existing_collection_ids = HashSet::from(["cid-user".to_string()]);
+        let changed_capabilities = HashSet::new();
+
+        let replay_collections = collections_requiring_replay(
+            &effective_collections,
+            &collection_cids,
+            &existing_collection_ids,
+            &changed_capabilities,
+        );
+
+        assert!(replay_collections.is_empty());
     }
 
     #[test]
